more collect_fixtures.py script into its own file
[ci skip-rust]

[ci skip-build-wheels]
